moved types to schema
